Handle diffs in the worker struct
Use the generated work list to execute diffs.
