+++ client/commands/state.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/ocuroot/ocuroot/client/release"
 	"github.com/ocuroot/ocuroot/client/state"
+	"github.com/ocuroot/ocuroot/client/work"
 	"github.com/ocuroot/ocuroot/refs"
 	"github.com/ocuroot/ocuroot/refs/refstore"
 	"github.com/ocuroot/ocuroot/sdk"
@@ -104,13 +105,19 @@ var StateDiffCmd = &cobra.Command{
 
 		cmd.SilenceUsage = true
 
-		diffs, err := state.Diff(ctx, tc.State, tc.Intent)
+		worker := &work.Worker{
+			Tracker: tc,
+		}
+
+		diffs, err := worker.Diff(ctx, work.IndentifyWorkRequest{
+			GitFilter: work.GitFilterCurrentCommitOnly,
+		})
 		if err != nil {
 			return fmt.Errorf("failed to diff: %w", err)
 		}
 
 		for _, diff := range diffs {
-			fmt.Println(diff)
+			fmt.Println(diff.Ref.String())
 		}
 
 		return nil
