+++ client/state/apply.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"path"
+	"reflect"
 	"strings"
 
 	"github.com/charmbracelet/log"
@@ -275,4 +276,38 @@ func applyDeletedDeployIntent(ctx context.Context, ref refs.Ref, state, intent r
 	}
 
 	return nil
+}
+
+func compareDeployIntent(
+	ctx context.Context,
+	state, intent refstore.Store,
+	intentRef refs.Ref,
+	stateRef refs.Ref,
+) (bool, error) {
+	var (
+		intentContent models.Intent
+		stateContent  models.Intent
+	)
+	if err := intent.Get(ctx, intentRef.String(), &intentContent); err != nil {
+		if err == refstore.ErrRefNotFound {
+			return false, nil
+		}
+		return false, fmt.Errorf("failed to get intent content: %w", err)
+	}
+	if err := state.Get(ctx, stateRef.String(), &stateContent); err != nil {
+		if err == refstore.ErrRefNotFound {
+			return false, nil
+		}
+		return false, fmt.Errorf("failed to get state content: %w", err)
+	}
+
+	log.Debug("Comparing deploy", "intentRef", intentRef.String(), "stateRef", stateRef.String(), "intentContent", intentContent, "stateContent", stateContent)
+
+	if !reflect.DeepEqual(intentContent, stateContent) {
+		log.Debug("Not equal")
+		return false, nil
+	}
+
+	log.Debug("Equal")
+	return true, nil
 }
